wip: finalize refactoring of separate modules
